pynotifyIPL/requirements.txt
pyqt
requests
bs4
